Automations/Linkedin_Post/requirements.txt
functions-framework==3.*
google-generativeai
google-cloud-bigquery
requests